# Optimize PreComputeTupleConstructor