@cartermp 2021-03-14
This is technically a breaking change given the change in exception that's thrown, but I think that's a very minor thing given the perf benefits.